black-hat-python chapter-7 polish
- indent size 2 to 4
- rename directory name
- fix some wrongs
