+++ black-hat-python/.gitignore
@@ -7,3 +7,12 @@
 
 ## Directory-based project format:
 .idea/
+
+# Environments
+.env
+.venv
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
