+++ black-hat-python/chapter07/environment.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+import os
+
+
+def run(**args):
+    print("[*] In environment module.")
+    return str(os.environ)
+
+
+if __name__ == '__main__':
+    print(run())
