--- black-hat-python/chapter7/dirlister.py
@@ -1,12 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-import os
-
-
-def run(**args):
-
-  print("[*] In dirlister module")
-  files = os.listdir(".")
-
-  return str(files)
