--- black-hat-python/chapter7/environment.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-import os
-
-
-def run(**args):
-  print("[*] In environment module.")
-  return str(os.environ)
